+++ src/opencodecs/__init__.py
@@ -191,4 +191,4 @@ __all__ = [
     "CziWriter", "CziPyramidWriter",
 ]
 
-__version__ = "0.1.5"
+__version__ = "0.1.6"